Extract free functions from Type.Annotation
